+++ src/param.h
@@ -169,7 +169,7 @@ public:
     vector<string> VcfSampleID; // size=total sample #
     vector<string> VcfSampleID_test; // sample id for ni_test in order of X[i][j]
     vector<string> InputSampleID; //size = ni_total
-    vector<pair<long long int, double> > UcharTable;
+    vector<pair<int, double> > UcharTable;
     vector<double> SNPmean;
     
 	bool mode_silence;
